rename templates folder
