+++ src/serverSetup/nunjucks.js
@@ -9,8 +9,8 @@ export function setupNunjucks ({ app, dataSubjects }) {
 
   const nunjucksEnv = nunjucks.configure([
     'src/views',
-    'src/views/checkPages',
-    'src/views/endpointSubmissionFormPages',
+    'src/views/check',
+    'src/views/submit',
     'node_modules/govuk-frontend/dist/',
     'node_modules/@x-govuk/govuk-prototype-components/'
   ], {
