+++ src/views/check/confirmation.html
