+++ src/views/check/geometry-type.html
