+++ src/views/check/start.html
