+++ src/views/check/statusPage/checkingFileMacro.html
