+++ src/views/check/statusPage/fileCheckedMacro.html
