+++ src/views/check/upload-method.html
