+++ src/views/submit/dataset-details.html
